requirements.txt
ydata-profiling
pandas
psycopg2-binary
sqlalchemy
python-dotenv